feat: restructured the folders
