+++ mava/systems/ppo/anakin/ff_ippo.py
@@ -30,20 +30,17 @@ from optax._src.base import OptState
 from rich.pretty import pprint
 
 from mava.evaluator import make_eval_fns
-from mava.networks import RecurrentActor as Actor
-from mava.networks import RecurrentValueNet as Critic
-from mava.networks import ScannedRNN
-from mava.systems.anakin.ppo.types import (
-    HiddenStates,
-    OptStates,
-    Params,
-    RNNLearnerState,
-    RNNPPOTransition,
-)
-from mava.types import ExperimentOutput, LearnerFn, RecActorApply, RecCriticApply
+from mava.networks import FeedForwardActor as Actor
+from mava.networks import FeedForwardValueNet as Critic
+from mava.systems.ppo.types import LearnerState, OptStates, Params, PPOTransition
+from mava.types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from mava.utils import make_env as environments
 from mava.utils.checkpointing import Checkpointer
-from mava.utils.jax_utils import unreplicate_batch_dim, unreplicate_n_dims
+from mava.utils.jax_utils import (
+    merge_leading_dims,
+    unreplicate_batch_dim,
+    unreplicate_n_dims,
+)
 from mava.utils.logger import LogEvent, MavaLogger
 from mava.utils.total_timestep_checker import check_total_timesteps
 from mava.utils.training import make_learning_rate
@@ -52,16 +49,17 @@ from mava.wrappers.episode_metrics import get_final_step_metrics
 
 def get_learner_fn(
     env: Environment,
-    apply_fns: Tuple[RecActorApply, RecCriticApply],
+    apply_fns: Tuple[ActorApply, CriticApply],
     update_fns: Tuple[optax.TransformUpdateFn, optax.TransformUpdateFn],
     config: DictConfig,
-) -> LearnerFn[RNNLearnerState]:
+) -> LearnerFn[LearnerState]:
     """Get the learner function."""
 
+    # Get apply and update functions for actor and critic networks.
     actor_apply_fn, critic_apply_fn = apply_fns
     actor_update_fn, critic_update_fn = update_fns
 
-    def _update_step(learner_state: RNNLearnerState, _: Any) -> Tuple[RNNLearnerState, Tuple]:
+    def _update_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, Tuple]:
         """A single update of the network.
 
         This function steps the environment and records the trajectory batch for
@@ -76,77 +74,35 @@ def get_learner_fn(
                 - key (PRNGKey): The random number generator state.
                 - env_state (State): The environment state.
                 - last_timestep (TimeStep): The last timestep in the current trajectory.
-                - dones (bool): Whether the last timestep was a terminal state.
-                - hstates (HiddenStates): The current hidden states of the RNN.
             _ (Any): The current metrics info.
         """
 
-        def _env_step(
-            learner_state: RNNLearnerState, _: Any
-        ) -> Tuple[RNNLearnerState, RNNPPOTransition]:
+        def _env_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, PPOTransition]:
             """Step the environment."""
-            (
-                params,
-                opt_states,
-                key,
-                env_state,
-                last_timestep,
-                last_done,
-                last_hstates,
-            ) = learner_state
+            params, opt_states, key, env_state, last_timestep = learner_state
 
+            # SELECT ACTION
             key, policy_key = jax.random.split(key)
+            actor_policy = actor_apply_fn(params.actor_params, last_timestep.observation)
+            value = critic_apply_fn(params.critic_params, last_timestep.observation)
 
-            # Add a batch dimension to the observation.
-            batched_observation = jax.tree_util.tree_map(
-                lambda x: x[jnp.newaxis, :], last_timestep.observation
-            )
-            ac_in = (
-                batched_observation,
-                last_done[jnp.newaxis, :],
-            )
-
-            # Run the network.
-            policy_hidden_state, actor_policy = actor_apply_fn(
-                params.actor_params, last_hstates.policy_hidden_state, ac_in
-            )
-            critic_hidden_state, value = critic_apply_fn(
-                params.critic_params, last_hstates.critic_hidden_state, ac_in
-            )
-
-            # Sample action from the policy and squeeze out the batch dimension.
             action = actor_policy.sample(seed=policy_key)
             log_prob = actor_policy.log_prob(action)
-            value, action, log_prob = (
-                value.squeeze(0),
-                action.squeeze(0),
-                log_prob.squeeze(0),
-            )
 
-            # Step the environment.
+            # STEP ENVIRONMENT
             env_state, timestep = jax.vmap(env.step, in_axes=(0, 0))(env_state, action)
 
-            # log episode return and length
+            # LOG EPISODE METRICS
             done = jax.tree_util.tree_map(
                 lambda x: jnp.repeat(x, config.system.num_agents).reshape(config.arch.num_envs, -1),
                 timestep.last(),
             )
             info = timestep.extras["episode_metrics"]
 
-            hstates = HiddenStates(policy_hidden_state, critic_hidden_state)
-            transition = RNNPPOTransition(
-                last_done,
-                action,
-                value,
-                timestep.reward,
-                log_prob,
-                last_timestep.observation,
-                last_hstates,
-                info,
+            transition = PPOTransition(
+                done, action, value, timestep.reward, log_prob, last_timestep.observation, info
             )
-            learner_state = RNNLearnerState(
-                params, opt_states, key, env_state, timestep, done, hstates
-            )
+            learner_state = LearnerState(params, opt_states, key, env_state, timestep)
             return learner_state, transition
 
         # STEP ENVIRONMENT FOR ROLLOUT LENGTH
@@ -155,53 +111,37 @@ def get_learner_fn(
         )
 
         # CALCULATE ADVANTAGE
-        (
-            params,
-            opt_states,
-            key,
-            env_state,
-            last_timestep,
-            last_done,
-            hstates,
-        ) = learner_state
-
-        # Add a batch dimension to the observation.
-        batched_last_observation = jax.tree_util.tree_map(
-            lambda x: x[jnp.newaxis, :], last_timestep.observation
-        )
-        ac_in = (
-            batched_last_observation,
-            last_done[jnp.newaxis, :],
-        )
-
-        # Run the network.
-        _, last_val = critic_apply_fn(params.critic_params, hstates.critic_hidden_state, ac_in)
-        # Squeeze out the batch dimension and mask out the value of terminal states.
-        last_val = last_val.squeeze(0)
+        params, opt_states, key, env_state, last_timestep = learner_state
+        last_val = critic_apply_fn(params.critic_params, last_timestep.observation)
 
         def _calculate_gae(
-            traj_batch: RNNPPOTransition, last_val: chex.Array, last_done: chex.Array
+            traj_batch: PPOTransition, last_val: chex.Array
         ) -> Tuple[chex.Array, chex.Array]:
-            def _get_advantages(
-                carry: Tuple[chex.Array, chex.Array, chex.Array], transition: RNNPPOTransition
-            ) -> Tuple[Tuple[chex.Array, chex.Array, chex.Array], chex.Array]:
-                gae, next_value, next_done = carry
-                done, value, reward = transition.done, transition.value, transition.reward
+            """Calculate the GAE."""
+
+            def _get_advantages(gae_and_next_value: Tuple, transition: PPOTransition) -> Tuple:
+                """Calculate the GAE for a single transition."""
+                gae, next_value = gae_and_next_value
+                done, value, reward = (
+                    transition.done,
+                    transition.value,
+                    transition.reward,
+                )
                 gamma = config.system.gamma
-                delta = reward + gamma * next_value * (1 - next_done) - value
-                gae = delta + gamma * config.system.gae_lambda * (1 - next_done) * gae
-                return (gae, value, done), gae
+                delta = reward + gamma * next_value * (1 - done) - value
+                gae = delta + gamma * config.system.gae_lambda * (1 - done) * gae
+                return (gae, value), gae
 
             _, advantages = jax.lax.scan(
                 _get_advantages,
-                (jnp.zeros_like(last_val), last_val, last_done),
+                (jnp.zeros_like(last_val), last_val),
                 traj_batch,
                 reverse=True,
                 unroll=16,
             )
             return advantages, advantages + traj_batch.value
 
-        advantages, targets = _calculate_gae(traj_batch, last_val, last_done)
+        advantages, targets = _calculate_gae(traj_batch, last_val)
 
         def _update_epoch(update_state: Tuple, _: Any) -> Tuple:
             """Update the network for a single epoch."""
@@ -216,19 +156,16 @@ def get_learner_fn(
                 def _actor_loss_fn(
                     actor_params: FrozenDict,
                     actor_opt_state: OptState,
-                    traj_batch: RNNPPOTransition,
+                    traj_batch: PPOTransition,
                     gae: chex.Array,
                     key: chex.PRNGKey,
                 ) -> Tuple:
                     """Calculate the actor loss."""
                     # RERUN NETWORK
-
-                    obs_and_done = (traj_batch.obs, traj_batch.done)
-                    _, actor_policy = actor_apply_fn(
-                        actor_params, traj_batch.hstates.policy_hidden_state[0], obs_and_done
-                    )
+                    actor_policy = actor_apply_fn(actor_params, traj_batch.obs)
                     log_prob = actor_policy.log_prob(traj_batch.action)
 
+                    # CALCULATE ACTOR LOSS
                     ratio = jnp.exp(log_prob - traj_batch.log_prob)
                     gae = (gae - gae.mean()) / (gae.std() + 1e-8)
                     loss_actor1 = ratio * gae
@@ -245,21 +182,18 @@ def get_learner_fn(
                     # The seed will be used in the TanhTransformedDistribution:
                     entropy = actor_policy.entropy(seed=key).mean()
 
-                    total_loss = loss_actor - config.system.ent_coef * entropy
-                    return total_loss, (loss_actor, entropy)
+                    total_loss_actor = loss_actor - config.system.ent_coef * entropy
+                    return total_loss_actor, (loss_actor, entropy)
 
                 def _critic_loss_fn(
                     critic_params: FrozenDict,
                     critic_opt_state: OptState,
-                    traj_batch: RNNPPOTransition,
+                    traj_batch: PPOTransition,
                     targets: chex.Array,
                 ) -> Tuple:
                     """Calculate the critic loss."""
                     # RERUN NETWORK
-                    obs_and_done = (traj_batch.obs, traj_batch.done)
-                    _, value = critic_apply_fn(
-                        critic_params, traj_batch.hstates.critic_hidden_state[0], obs_and_done
-                    )
+                    value = critic_apply_fn(critic_params, traj_batch.obs)
 
                     # CALCULATE VALUE LOSS
                     value_pred_clipped = traj_batch.value + (value - traj_batch.value).clip(
@@ -269,8 +203,8 @@ def get_learner_fn(
                     value_losses_clipped = jnp.square(value_pred_clipped - targets)
                     value_loss = 0.5 * jnp.maximum(value_losses, value_losses_clipped).mean()
 
-                    total_loss = config.system.vf_coef * value_loss
-                    return total_loss, (value_loss)
+                    critic_total_loss = config.system.vf_coef * value_loss
+                    return critic_total_loss, (value_loss)
 
                 # CALCULATE ACTOR LOSS
                 key, entropy_key = jax.random.split(key)
@@ -321,6 +255,7 @@ def get_learner_fn(
                 )
                 critic_new_params = optax.apply_updates(params.critic_params, critic_updates)
 
+                # PACK NEW PARAMS AND OPTIMISER STATE
                 new_params = Params(actor_new_params, critic_new_params)
                 new_opt_state = OptStates(actor_new_opt_state, critic_new_opt_state)
 
@@ -335,62 +270,33 @@ def get_learner_fn(
                     "actor_loss": actor_loss,
                     "entropy": entropy,
                 }
-
                 return (new_params, new_opt_state, entropy_key), loss_info
 
             params, opt_states, traj_batch, advantages, targets, key = update_state
             key, shuffle_key, entropy_key = jax.random.split(key, 3)
 
             # SHUFFLE MINIBATCHES
+            batch_size = config.system.rollout_length * config.arch.num_envs
+            permutation = jax.random.permutation(shuffle_key, batch_size)
             batch = (traj_batch, advantages, targets)
-            num_recurrent_chunks = (
-                config.system.rollout_length // config.system.recurrent_chunk_size
-            )
-            batch = jax.tree_util.tree_map(
-                lambda x: x.reshape(
-                    config.system.recurrent_chunk_size,
-                    config.arch.num_envs * num_recurrent_chunks,
-                    *x.shape[2:],
-                ),
-                batch,
-            )
-            permutation = jax.random.permutation(
-                shuffle_key, config.arch.num_envs * num_recurrent_chunks
-            )
+            batch = jax.tree_util.tree_map(lambda x: merge_leading_dims(x, 2), batch)
             shuffled_batch = jax.tree_util.tree_map(
-                lambda x: jnp.take(x, permutation, axis=1), batch
+                lambda x: jnp.take(x, permutation, axis=0), batch
             )
-            reshaped_batch = jax.tree_util.tree_map(
-                lambda x: jnp.reshape(
-                    x, (x.shape[0], config.system.num_minibatches, -1, *x.shape[2:])
-                ),
+            minibatches = jax.tree_util.tree_map(
+                lambda x: jnp.reshape(x, [config.system.num_minibatches, -1] + list(x.shape[1:])),
                 shuffled_batch,
             )
-            minibatches = jax.tree_util.tree_map(lambda x: jnp.swapaxes(x, 1, 0), reshaped_batch)
 
             # UPDATE MINIBATCHES
             (params, opt_states, entropy_key), loss_info = jax.lax.scan(
                 _update_minibatch, (params, opt_states, entropy_key), minibatches
             )
 
-            update_state = (
-                params,
-                opt_states,
-                traj_batch,
-                advantages,
-                targets,
-                key,
-            )
+            update_state = (params, opt_states, traj_batch, advantages, targets, key)
             return update_state, loss_info
 
-        update_state = (
-            params,
-            opt_states,
-            traj_batch,
-            advantages,
-            targets,
-            key,
-        )
+        update_state = (params, opt_states, traj_batch, advantages, targets, key)
 
         # UPDATE EPOCHS
         update_state, loss_info = jax.lax.scan(
@@ -398,19 +304,11 @@ def get_learner_fn(
         )
 
         params, opt_states, traj_batch, advantages, targets, key = update_state
-        learner_state = RNNLearnerState(
-            params,
-            opt_states,
-            key,
-            env_state,
-            last_timestep,
-            last_done,
-            hstates,
-        )
+        learner_state = LearnerState(params, opt_states, key, env_state, last_timestep)
         metric = traj_batch.info
         return learner_state, (metric, loss_info)
 
-    def learner_fn(learner_state: RNNLearnerState) -> ExperimentOutput[RNNLearnerState]:
+    def learner_fn(learner_state: LearnerState) -> ExperimentOutput[LearnerState]:
         """Learner function.
 
         This function represents the learner, it updates the network parameters
@@ -420,12 +318,10 @@ def get_learner_fn(
         Args:
             learner_state (NamedTuple):
                 - params (Params): The initial model parameters.
-                - opt_states (OptStates): The initial optimizer states.
+                - opt_states (OptStates): The initial optimizer state.
                 - key (chex.PRNGKey): The random number generator state.
                 - env_state (LogEnvState): The environment state.
                 - timesteps (TimeStep): The initial timestep in the initial trajectory.
-                - dones (bool): Whether the initial timestep was a terminal state.
-                - hstateS (HiddenStates): The initial hidden states of the RNN.
         """
 
         batched_update_step = jax.vmap(_update_step, in_axes=(0, None), axis_name="batch")
@@ -444,38 +340,26 @@ def get_learner_fn(
 
 def learner_setup(
     env: Environment, keys: chex.Array, config: DictConfig
-) -> Tuple[LearnerFn[RNNLearnerState], Actor, RNNLearnerState]:
+) -> Tuple[LearnerFn[LearnerState], Actor, LearnerState]:
     """Initialise learner_fn, network, optimiser, environment and states."""
     # Get available TPU cores.
     n_devices = len(jax.devices())
 
     # Get number of agents.
-    num_agents = env.num_agents
-    config.system.num_agents = num_agents
+    config.system.num_agents = env.num_agents
 
     # PRNG keys.
     key, actor_net_key, critic_net_key = keys
 
-    # Define network and optimisers.
-    actor_pre_torso = hydra.utils.instantiate(config.network.actor_network.pre_torso)
-    actor_post_torso = hydra.utils.instantiate(config.network.actor_network.post_torso)
+    # Define network and optimiser.
+    actor_torso = hydra.utils.instantiate(config.network.actor_network.pre_torso)
     actor_action_head = hydra.utils.instantiate(
         config.network.action_head, action_dim=env.action_dim
     )
-    critic_pre_torso = hydra.utils.instantiate(config.network.critic_network.pre_torso)
-    critic_post_torso = hydra.utils.instantiate(config.network.critic_network.post_torso)
+    critic_torso = hydra.utils.instantiate(config.network.critic_network.pre_torso)
 
-    actor_network = Actor(
-        pre_torso=actor_pre_torso,
-        post_torso=actor_post_torso,
-        action_head=actor_action_head,
-        hidden_state_dim=config.network.hidden_state_dim,
-    )
-    critic_network = Critic(
-        pre_torso=critic_pre_torso,
-        post_torso=critic_post_torso,
-        hidden_state_dim=config.network.hidden_state_dim,
-    )
+    actor_network = Actor(torso=actor_torso, action_head=actor_action_head)
+    critic_network = Critic(torso=critic_torso)
 
     actor_lr = make_learning_rate(config.system.actor_lr, config)
     critic_lr = make_learning_rate(config.system.critic_lr, config)
@@ -490,54 +374,27 @@ def learner_setup(
     )
 
     # Initialise observation with obs of all agents.
-    init_obs = env.observation_spec().generate_value()
-    init_obs = jax.tree_util.tree_map(
-        lambda x: jnp.repeat(x[jnp.newaxis, ...], config.arch.num_envs, axis=0),
-        init_obs,
-    )
-    init_obs = jax.tree_util.tree_map(lambda x: x[jnp.newaxis, ...], init_obs)
-    init_done = jnp.zeros((1, config.arch.num_envs, num_agents), dtype=bool)
-    init_x = (init_obs, init_done)
+    obs = env.observation_spec().generate_value()
+    init_x = jax.tree_util.tree_map(lambda x: x[jnp.newaxis, ...], obs)
 
-    # Initialise hidden states.
-    init_policy_hstate = ScannedRNN.initialize_carry(
-        (config.arch.num_envs, num_agents), config.network.hidden_state_dim
-    )
-    init_critic_hstate = ScannedRNN.initialize_carry(
-        (config.arch.num_envs, num_agents), config.network.hidden_state_dim
-    )
-
-    # initialise params and optimiser state.
-    actor_params = actor_network.init(actor_net_key, init_policy_hstate, init_x)
+    # Initialise actor params and optimiser state.
+    actor_params = actor_network.init(actor_net_key, init_x)
     actor_opt_state = actor_optim.init(actor_params)
-    critic_params = critic_network.init(critic_net_key, init_critic_hstate, init_x)
+
+    # Initialise critic params and optimiser state.
+    critic_params = critic_network.init(critic_net_key, init_x)
     critic_opt_state = critic_optim.init(critic_params)
 
-    # Get network apply functions and optimiser updates.
+    # Pack params.
+    params = Params(actor_params, critic_params)
+
+    # Pack apply and update functions.
     apply_fns = (actor_network.apply, critic_network.apply)
     update_fns = (actor_optim.update, critic_optim.update)
 
     # Get batched iterated update and replicate it to pmap it over cores.
     learn = get_learner_fn(env, apply_fns, update_fns, config)
     learn = jax.pmap(learn, axis_name="device")
-
-    # Pack params and initial states.
-    params = Params(actor_params, critic_params)
-    hstates = HiddenStates(init_policy_hstate, init_critic_hstate)
-
-    # Load model from checkpoint if specified.
-    if config.logger.checkpointing.load_model:
-        loaded_checkpoint = Checkpointer(
-            model_name=config.logger.system_name,
-            **config.logger.checkpointing.load_args,  # Other checkpoint args
-        )
-        # Restore the learner state from the checkpoint
-        restored_params, restored_hstates = loaded_checkpoint.restore_params(
-            input_params=params, restore_hstates=True, THiddenState=HiddenStates
-        )
-        # Update the params and hstates
-        params = restored_params
-        hstates = restored_hstates if restored_hstates else hstates
 
     # Initialise environment states and timesteps: across devices and batches.
     key, *env_keys = jax.random.split(
@@ -553,14 +410,21 @@ def learner_setup(
     env_states = jax.tree_map(reshape_states, env_states)
     timesteps = jax.tree_map(reshape_states, timesteps)
 
+    # Load model from checkpoint if specified.
+    if config.logger.checkpointing.load_model:
+        loaded_checkpoint = Checkpointer(
+            model_name=config.logger.system_name,
+            **config.logger.checkpointing.load_args,  # Other checkpoint args
+        )
+        # Restore the learner state from the checkpoint
+        restored_params, _ = loaded_checkpoint.restore_params(input_params=params)
+        # Update the params
+        params = restored_params
+
     # Define params to be replicated across devices and batches.
-    dones = jnp.zeros(
-        (config.arch.num_envs, num_agents),
-        dtype=bool,
-    )
     key, step_keys = jax.random.split(key)
     opt_states = OptStates(actor_opt_state, critic_opt_state)
-    replicate_learner = (params, opt_states, hstates, step_keys, dones)
+    replicate_learner = (params, opt_states, step_keys)
 
     # Duplicate learner for update_batch_size.
     broadcast = lambda x: jnp.broadcast_to(x, (config.system.update_batch_size,) + x.shape)
@@ -570,32 +434,17 @@ def learner_setup(
     replicate_learner = flax.jax_utils.replicate(replicate_learner, devices=jax.devices())
 
     # Initialise learner state.
-    params, opt_states, hstates, step_keys, dones = replicate_learner
-    init_learner_state = RNNLearnerState(
-        params=params,
-        opt_states=opt_states,
-        key=step_keys,
-        env_state=env_states,
-        timestep=timesteps,
-        dones=dones,
-        hstates=hstates,
-    )
+    params, opt_states, step_keys = replicate_learner
+    init_learner_state = LearnerState(params, opt_states, step_keys, env_states, timesteps)
+
     return learn, actor_network, init_learner_state
 
 
-def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
+def run_experiment(_config: DictConfig) -> float:
     """Runs experiment."""
     config = copy.deepcopy(_config)
 
     n_devices = len(jax.devices())
-
-    # Set recurrent chunk size.
-    if config.system.recurrent_chunk_size is None:
-        config.system.recurrent_chunk_size = config.system.rollout_length
-    else:
-        assert (
-            config.system.rollout_length % config.system.recurrent_chunk_size == 0
-        ), "Rollout length must be divisible by recurrent chunk size."
 
     # Create the enviroments for train and eval.
     env, eval_env = environments.make(config)
@@ -613,13 +462,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
     # Setup evaluator.
     # One key per device for evaluation.
     eval_keys = jax.random.split(key_e, n_devices)
-    evaluator, absolute_metric_evaluator = make_eval_fns(
-        eval_env=eval_env,
-        network_apply_fn=actor_network.apply,
-        config=config,
-        use_recurrent_net=True,
-        scanned_rnn=ScannedRNN,
-    )
+    evaluator, absolute_metric_evaluator = make_eval_fns(eval_env, actor_network.apply, config)
 
     # Calculate total timesteps.
     config = check_total_timesteps(config)
@@ -658,6 +501,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
     for eval_step in range(config.arch.num_evaluation):
         # Train.
         start_time = time.time()
+
         learner_output = learn(learner_state)
         jax.block_until_ready(learner_output)
 
@@ -723,7 +567,6 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
         jax.block_until_ready(evaluator_output)
 
         elapsed_time = time.time() - start_time
-
         steps_per_eval = int(jnp.sum(evaluator_output.episode_metrics["episode_length"]))
         t = int(steps_per_rollout * (eval_step + 1))
         evaluator_output.episode_metrics["steps_per_second"] = steps_per_eval / elapsed_time
@@ -735,7 +578,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
     return eval_performance
 
 
-@hydra.main(config_path="../../../configs", config_name="default_rec_ippo.yaml", version_base="1.2")
+@hydra.main(config_path="../../../configs", config_name="default_ff_ippo.yaml", version_base="1.2")
 def hydra_entry_point(cfg: DictConfig) -> float:
     """Experiment entry point."""
     # Allow dynamic attributes.
@@ -743,7 +586,7 @@ def hydra_entry_point(cfg: DictConfig) -> float:
 
     # Run experiment.
     eval_performance = run_experiment(cfg)
-    print(f"{Fore.CYAN}{Style.BRIGHT}Recurrent IPPO experiment completed{Style.RESET_ALL}")
+    print(f"{Fore.CYAN}{Style.BRIGHT}IPPO experiment completed{Style.RESET_ALL}")
     return eval_performance
