+++ mava/systems/ppo/anakin/rec_ippo.py
@@ -33,7 +33,7 @@ from mava.evaluator import make_eval_fns
 from mava.networks import RecurrentActor as Actor
 from mava.networks import RecurrentValueNet as Critic
 from mava.networks import ScannedRNN
-from mava.systems.anakin.ppo.types import (
+from mava.systems.ppo.types import (
     HiddenStates,
     OptStates,
     Params,
@@ -76,8 +76,8 @@ def get_learner_fn(
                 - key (PRNGKey): The random number generator state.
                 - env_state (State): The environment state.
                 - last_timestep (TimeStep): The last timestep in the current trajectory.
-                - last_done (bool): Whether the last timestep was a terminal state.
-                - hstates (HiddenStates): The hidden state of the policy and critic RNN.
+                - dones (bool): Whether the last timestep was a terminal state.
+                - hstates (HiddenStates): The current hidden states of the RNN.
             _ (Any): The current metrics info.
         """
 
@@ -101,7 +101,10 @@ def get_learner_fn(
             batched_observation = jax.tree_util.tree_map(
                 lambda x: x[jnp.newaxis, :], last_timestep.observation
             )
-            ac_in = (batched_observation, last_done[jnp.newaxis, :])
+            ac_in = (
+                batched_observation,
+                last_done[jnp.newaxis, :],
+            )
 
             # Run the network.
             policy_hidden_state, actor_policy = actor_apply_fn(
@@ -114,8 +117,11 @@ def get_learner_fn(
             # Sample action from the policy and squeeze out the batch dimension.
             action = actor_policy.sample(seed=policy_key)
             log_prob = actor_policy.log_prob(action)
-
-            action, log_prob, value = (action.squeeze(0), log_prob.squeeze(0), value.squeeze(0))
+            value, action, log_prob = (
+                value.squeeze(0),
+                action.squeeze(0),
+                log_prob.squeeze(0),
+            )
 
             # Step the environment.
             env_state, timestep = jax.vmap(env.step, in_axes=(0, 0))(env_state, action)
@@ -163,11 +169,13 @@ def get_learner_fn(
         batched_last_observation = jax.tree_util.tree_map(
             lambda x: x[jnp.newaxis, :], last_timestep.observation
         )
-        ac_in = (batched_last_observation, last_done[jnp.newaxis, :])
+        ac_in = (
+            batched_last_observation,
+            last_done[jnp.newaxis, :],
+        )
 
         # Run the network.
         _, last_val = critic_apply_fn(params.critic_params, hstates.critic_hidden_state, ac_in)
-
         # Squeeze out the batch dimension and mask out the value of terminal states.
         last_val = last_val.squeeze(0)
 
@@ -214,6 +222,7 @@ def get_learner_fn(
                 ) -> Tuple:
                     """Calculate the actor loss."""
                     # RERUN NETWORK
+
                     obs_and_done = (traj_batch.obs, traj_batch.done)
                     _, actor_policy = actor_apply_fn(
                         actor_params, traj_batch.hstates.policy_hidden_state[0], obs_and_done
@@ -416,7 +425,7 @@ def get_learner_fn(
                 - env_state (LogEnvState): The environment state.
                 - timesteps (TimeStep): The initial timestep in the initial trajectory.
                 - dones (bool): Whether the initial timestep was a terminal state.
-                - hstates (HiddenStates): The hidden state of the policy and critic RNN.
+                - hstateS (HiddenStates): The initial hidden states of the RNN.
         """
 
         batched_update_step = jax.vmap(_update_step, in_axes=(0, None), axis_name="batch")
@@ -447,7 +456,7 @@ def learner_setup(
     # PRNG keys.
     key, actor_net_key, critic_net_key = keys
 
-    # Define network and optimiser.
+    # Define network and optimisers.
     actor_pre_torso = hydra.utils.instantiate(config.network.actor_network.pre_torso)
     actor_post_torso = hydra.utils.instantiate(config.network.actor_network.post_torso)
     actor_action_head = hydra.utils.instantiate(
@@ -466,7 +475,6 @@ def learner_setup(
         pre_torso=critic_pre_torso,
         post_torso=critic_post_torso,
         hidden_state_dim=config.network.hidden_state_dim,
-        centralised_critic=True,
     )
 
     actor_lr = make_learning_rate(config.system.actor_lr, config)
@@ -489,9 +497,9 @@ def learner_setup(
     )
     init_obs = jax.tree_util.tree_map(lambda x: x[jnp.newaxis, ...], init_obs)
     init_done = jnp.zeros((1, config.arch.num_envs, num_agents), dtype=bool)
-    init_obs_done = (init_obs, init_done)
+    init_x = (init_obs, init_done)
 
-    # Initialise hidden state.
+    # Initialise hidden states.
     init_policy_hstate = ScannedRNN.initialize_carry(
         (config.arch.num_envs, num_agents), config.network.hidden_state_dim
     )
@@ -500,9 +508,9 @@ def learner_setup(
     )
 
     # initialise params and optimiser state.
-    actor_params = actor_network.init(actor_net_key, init_policy_hstate, init_obs_done)
+    actor_params = actor_network.init(actor_net_key, init_policy_hstate, init_x)
     actor_opt_state = actor_optim.init(actor_params)
-    critic_params = critic_network.init(critic_net_key, init_critic_hstate, init_obs_done)
+    critic_params = critic_network.init(critic_net_key, init_critic_hstate, init_x)
     critic_opt_state = critic_optim.init(critic_params)
 
     # Get network apply functions and optimiser updates.
@@ -590,7 +598,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
         ), "Rollout length must be divisible by recurrent chunk size."
 
     # Create the enviroments for train and eval.
-    env, eval_env = environments.make(config=config, add_global_state=True)
+    env, eval_env = environments.make(config)
 
     # PRNG keys.
     key, key_e, actor_net_key, critic_net_key = jax.random.split(
@@ -628,6 +636,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
         * config.system.update_batch_size
         * config.arch.num_envs
     )
+
     # Logger setup
     logger = MavaLogger(config)
     cfg: Dict = OmegaConf.to_container(config, resolve=True)
@@ -726,9 +735,7 @@ def run_experiment(_config: DictConfig) -> float:  # noqa: CCR001
     return eval_performance
 
 
-@hydra.main(
-    config_path="../../../configs", config_name="default_rec_mappo.yaml", version_base="1.2"
-)
+@hydra.main(config_path="../../../configs", config_name="default_rec_ippo.yaml", version_base="1.2")
 def hydra_entry_point(cfg: DictConfig) -> float:
     """Experiment entry point."""
     # Allow dynamic attributes.
@@ -736,7 +743,7 @@ def hydra_entry_point(cfg: DictConfig) -> float:
 
     # Run experiment.
     eval_performance = run_experiment(cfg)
-    print(f"{Fore.CYAN}{Style.BRIGHT}Recurrent MAPPO experiment completed{Style.RESET_ALL}")
+    print(f"{Fore.CYAN}{Style.BRIGHT}Recurrent IPPO experiment completed{Style.RESET_ALL}")
     return eval_performance
